Fix no signature of method

+++ 2026/data/grack_exhibition/grack/script/finish_parkour.groovy
@@ -43,7 +43,7 @@ if (source.isPlayer()) {
                 if (s.length() == 9) {
                     boolean success = true
                     for (char c : s.toCharArray()) {
-                        if (ch.indexOf(c) == -1) {
+                        if (ch.indexOf((int) c) == -1) {
                             success = false
                             break
                         }
